Crate impl_more

Concise, declarative trait implementation macros.

#[no_std]

Where possible, these macros emit #[no_std]-compatible code.

Macros

forward_deref_and_mut

Implements Deref and DerefMut by forwarding through an inner field’s implementation.

forward_display

Implements Display for structs by forwarding to one of its field.

forward_error

Implements Error for structs and forwards the source implementation to one of its fields.

impl_as_mut

Implement AsMut for a struct.

impl_as_ref

Implement AsRef for a struct.

impl_deref

Implement Deref for a struct.

impl_deref_and_mut

Implements Deref and DerefMut by forwarding through an inner field’s implementation.

impl_deref_mut

Implement DerefMut for a struct.

impl_display

Implements Display for structs using a format!-like string constructor.

impl_display_enum

Implements Display for enums using a static string or format args for each variant.

impl_error_enum

Implements Error for enums.

impl_from

Implement From for a struct.

impl_from_for_primitive

Implement From for a primitive.

impl_into

Implement Into for a struct.